+++ cpp_impl/perf.h
@@ -1,6 +1,6 @@
 #include <functional> // <-- Add this
 
-using comp_func = std::function<void(float *X, float *B, float *Y, int M, int N, int K)>;
+using comp_func = std::function<void(double *X, double *B, double *Y, int M, int N, int K)>;
 
 // perf_test still generates data based on M, K, N, nonZero,
 // but the comp_func 'f' it receives already has its specific sparse matrix data captured.
